+++ .github/workflows/cicd-main.yml
@@ -309,15 +309,6 @@ jobs:
        SCRIPT: |
          CUDA_VISIBLE_DEVICES="" NEMO_NUMBA_MINVER=0.53 pytest tests/lightning -m "not pleasefixme" --cpu --with_downloads --relax_numba_compat
 
-  L0_Unit_Tests_CPU_Utils:
-     needs: [cicd-test-container-setup]
-     uses: ./.github/workflows/_test_template.yml
-     if: contains(fromJSON(needs.cicd-test-container-setup.outputs.test_to_run), 'L0_Unit_Tests_CPU_Utils') || needs.cicd-test-container-setup.outputs.all == 'true'
-     with:
-       RUNNER: self-hosted-azure-cpu
-       SCRIPT: |
-         CUDA_VISIBLE_DEVICES="" NEMO_NUMBA_MINVER=0.53 pytest tests/utils -m "not pleasefixme" --cpu --with_downloads --relax_numba_compat
-
   L0_Unit_Tests_CPU_Others:
      needs: [cicd-test-container-setup]
      uses: ./.github/workflows/_test_template.yml
